+++ lib/shipcloud/pickup_request.rb
@@ -2,7 +2,7 @@ module Shipcloud
   class PickupRequest < Base
     include Shipcloud::Operations::Create
 
-    attr_accessor :carrier, :pickup_time, :shipments
+    attr_accessor :carrier, :pickup_time, :shipments, :pickup_address
     attr_reader :id, :carrier_pickup_number
 
     def self.base_url