+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 ### Fixed
+- Parse response only when it is not empty
 
 ## [0.7.0] - 2016-01-21
 ### Added