+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [0.8.0] - 2017-07-03
+### Added
 - Add attribute ```metadata``` to class ```Shipcloud::Shipment``` in order to transmit JSON data (#16).
 - Add resource pickup_request in order to submit pickup request to shipcloud
 - Add attribute ```pickup_address``` to class ```Shipcloud::PickupRequest``` to submit an alternative address for pickup request to shipcloud
@@ -7,10 +20,6 @@
 - Add attribute ```deactivated``` to class ```Shipcloud::Webhook```
 - Add ```affiliate_id``` to ```Shipcloud::Configuration``` and submit it (or a default affiliate id) via API headers to shipcloud
 - Add class ```Shipcloud::Tracker``` with create, find, and index operations
-
-### Removed
-
-### Changed
 
 ### Fixed
 - Parse response only when it is not empty